refactor: 견스타그램 관련 컴포넌트 재사용성을 위한 컴포넌트 분리
